+++ lib/badgeville/errors.rb
@@ -1,0 +1,31 @@
+module Badgeville
+  class Error < StandardError
+    attr_accessor :data
+
+    def initialize(data)
+      @data = data
+    end
+  end
+
+  # deprecated old error
+  class BadgevilleError < Error
+    attr_accessor :code
+
+    def initialize(error_code = nil, data = nil)
+      super(data.to_s)
+      @code = error_code
+    end
+  end
+
+  # Timeout
+  class NotAvailable < Error; end
+
+  # 404
+  class NotFound < Error; end
+
+  # 500
+  class ServerError < Error; end
+
+  # invalid JSON returned etc.
+  class ParseError < Error; end
+end